toolkit/tools/stakeholder_briefing/requirements.txt
openai>=1.12.0
python-docx>=1.1.0
gnews>=0.3.0
requests>=2.28.0
python-dateutil>=2.8.0
